pyproject.toml
[tool.poetry]
name = "pg-loginator"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "~3.9"
psycopg2-binary = "^2.9.9"
paramiko = "^3.4.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
tox = "^4.14.2"
pytest = "^8.1.1"
pytest-cov = "^5.0.0"

[tool.pytest.ini_options]
addopts = "--cov=pg-logidater"

[build-system]
requires = ["setuptools>=36.6.0", "wheel", 'pbr']

